Updated paseExpression and added parseTerm as according to Ashhad to handle expressions in Evaluator.
