+++ Parser.cpp
@@ -15,6 +15,7 @@ using namespace std;
  *
  * @return An Abstract Syntax Tree root node representing the program.
  */
+
 ASTNode Parser::parse()
 {
     ASTNode root("PROGRAM");
@@ -48,11 +49,12 @@ ASTNode Parser::parse()
  *
  * @return An Abstract Syntax Tree node representing the assignment statement.
  */
+
 ASTNode Parser::parseAssignment()
 {
     Token identifier = consume("IDENTIFIER");
     consume("ASSIGN");
-    
+
     ASTNode assignment("ASSIGNMENT");
     assignment.children.push_back(ASTNode("IDENTIFIER", identifier.value));
 
@@ -62,45 +64,8 @@ ASTNode Parser::parseAssignment()
 
     consume("SEMICOLON");
     consume("LINE_BREAK");
-    
+
     return assignment;
-}
-
-ASTNode Parser::parseExpression()
-{
-    ASTNode left = parseTerm();
-
-    while (peek().type == "ARITHMETIC_OPERATOR")
-    {
-        Token op = consume("ARITHMETIC_OPERATOR");
-        ASTNode right = parseTerm();
-
-        ASTNode expr("EXPRESSION");
-        expr.children.push_back(left);
-        expr.children.push_back(ASTNode("ARITHMETIC_OPERATOR", op.value));
-        expr.children.push_back(right);
-
-        left = expr;
-    }
-
-    return left;
-}
-
-ASTNode Parser::parseTerm()
-{
-    Token token = peek();
-    if (token.type == "INTEGER" || token.type == "FLOAT" || token.type == "DOUBLE")
-    {
-        return ASTNode(token.type, consume(token.type).value);
-    }
-    else if (token.type == "IDENTIFIER")
-    {
-        return ASTNode("IDENTIFIER", consume("IDENTIFIER").value);
-    }
-    else
-    {
-        throw runtime_error("Unexpected token in expression: " + token.value);
-    }
 }
 
 /**
@@ -149,6 +114,7 @@ ASTNode Parser::parsePrintStatement()
  *
  * @return An Abstract Syntax Tree node representing the if statement.
  */
+
 ASTNode Parser::parseIfStatement()
 {
     consume("CHECK_IF");
@@ -194,6 +160,7 @@ ASTNode Parser::parseIfStatement()
  *
  * @return An Abstract Syntax Tree node representing the while loop statement.
  */
+
 ASTNode Parser::parseWhileLoop()
 {
     consume("WHILE");
